+++ src/qnethernet/QNEthernetFrame.h
@@ -23,6 +23,7 @@
 #include "lwip/prot/ethernet.h"
 #include "lwip_driver.h"
 #include "qnethernet/StaticInit.h"
+#include "qnethernet/internal/CircularBuffer.h"
 #include "qnethernet/internal/PrintfChecked.h"
 #include "qnethernet/internal/optional.h"
 
@@ -190,12 +191,12 @@ class EthernetFrameClass final : public Stream, public internal::PrintfChecked {
 
   // Returns the receive queue capacity.
   size_t receiveQueueCapacity() const {
-    return inBuf_.size();
+    return inBuf_.capacity();
   }
 
   // Returns the number of frames currently in the receive queue.
   size_t receiveQueueSize() const {
-    return inBufSize_;
+    return inBuf_.size();
   }
 
   // Returns the total number of dropped received frames since reception was
@@ -240,10 +241,7 @@ class EthernetFrameClass final : public Stream, public internal::PrintfChecked {
   bool isAvailable() const;
 
   // Received frames; updated every time one is received
-  std::vector<Frame> inBuf_;  // Holds received frames
-  size_t inBufTail_ = 0;
-  size_t inBufHead_ = 0;
-  size_t inBufSize_ = 0;
+  internal::CircularBuffer<Frame> inBuf_;
 
   // Frame being processed by the caller
   Frame frame_;   // Holds the frame being read